third_party/blink/web_tests/fast/forms/label/labels-item-index-expected.txt

Test that the items of a labels list can be accessed individually.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS labels.item(1).id is "l11"
PASS labels.item(1).id is "l12"
PASS labels.item(1).id is "l13"
PASS labels.item(1).id is "l14"
PASS successfullyParsed is true

TEST COMPLETE
